+++ packages/ui/Spinner/Spinner.js
@@ -92,6 +92,15 @@ const Spinner = kind({
 		css: PropTypes.object,
 
 		/**
+		 * Halts the animation of the spinner
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		paused: PropTypes.bool,
+
+		/**
 		 * Sets a scrim behind the spinner with the `css.scrim` class applied.
 		 *
 		 * Only has an effect when `blockClickOn` is `'screen'` or `'container'` and has no effect
@@ -106,18 +115,19 @@ const Spinner = kind({
 
 	defaultProps: {
 		centered: false,
+		paused: false,
 		scrim: false
 	},
 
 	styles: {
 		css: componentCss,
-		className: 'spinner running',
+		className: 'spinner',
 		publicClassNames: true
 	},
 
 	computed: {
-		className: ({centered, styler}) => styler.append(
-			{centered}
+		className: ({centered, paused, styler}) => styler.append(
+			{centered, running: !paused}
 		),
 		scrimClassName: ({blockClickOn, scrim, styler}) => styler.join(
 			{blockClickOn, scrim}
@@ -130,6 +140,7 @@ const Spinner = kind({
 
 	render: ({blockClickOn, component: Component, scrimClassName, scrimType, spinnerContainerClassName, ...rest}) =>  {
 		delete rest.centered;
+		delete rest.paused;
 		delete rest.scrim;
 
 		switch (blockClickOn) {