+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/Layout` `ref` support so developers can safely get refs to their layout elements
+- `ui/Item`, `ui/Layout`, `ui/Repeater`, `ui/SlotItem`, `ui/Spinner`, `ui/ToggleItem`, and `ui/ViewManager` support for `ref` to gain access to the wrapped `component`
 - `ui/Repeater` prop `component` to allow customization of its base element
 - `ui/Spinner` prop `paused` to halt the animation. Previously this was hard-coded "on", but now it can be toggled.